sprint_2: feat: создал компоненту InputForm, рефакторинг страницы регистрации
